requirements.txt
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.1
idna==2.10
PyJWT==1.7.1
requests==2.25.1
urllib3==1.26.3
